Improved disk size IOCTL checks
Previously we were only checking PhysicalDrive0 with the size IOCTL check, but on some systems PhysicalDrive0 might be associated with a single disk in a software RAID volume or an appliance boot volume. Such cases would be incorrectly flagged as a VM for being under 80GB in size. This new code instead finds the OS drive letter, maps that to a set of physical disks, then sums the sizes.  If anything goes wrong it falls back to the old method of just checking PhysicalDrive0.