Improved/fixed ThreadHideFromDebugger checks
Previously this check didn't actually really check anything - it just tried to set the flag and returned true if the call to NtSetInformationThread didn't succeed. This wasn't a very good check so I've enhanced it significantly. We now feed NtSetInformationThread various bogus values to try to catch out hooks that ignore certain arguments, do the old check of seeing if the call succeeded, and then verify that the call actually set the flag.