Completed switching all LoadLibrary / GetProcAddress calls to the API class methods
Now using API::IsAvailable() and API::GetAPI() so that we can check all APIs for availability at the start of execution and save ourselves writing the same API definitions and lookups a bunch of times.

The only thing not converted to the new style is the hooking code and any checks that require direct use of GetProcAddress.