Patching int2d check to fix #223
The int 0x2d trick requires that rax be set to 0 or some other nontrivial value. If we don't explicitly set it here, we rely upon the compiler to just happen to not put 1, 3, 4, or 5 in rax when this trick is performed. We only got away with this up until now out of sheer luck!