Improve parent process check to avoid false positives.
The previous method of checking if the process was launched by Explorer was very prone to false positives, since it would return a detection if you have Explorer process isolation enabled (i.e. one process per explorer window).

The new method finds the parent process ID and checks the main module path to see if it's an explorer.exe process. The full path is checked so as to avoid the trick of just renaming the parent process to explorer.exe somewhere else in the filesystem.

If the parent process filename or explorer path cannot be fetched for some reason, the check falls back to the old method. It also handles the situation where GetExplorerPIDbyShellWindow fails and returns zero.